使用遮罩来实现单位的高亮

+++ Ryujin_SDL2/Bullet.h
@@ -22,7 +22,7 @@ namespace NspBullet {
 		//种类、计数器、颜色、状态、保证不消失的最短时间、效果的种类
 		int  knd, col, state, till, eff;
 		//坐标、速度、基本角度、瞬间记忆速度
-		double  spd, base_angle[1], rem_spd[1];
+		float  spd, base_angle[1], rem_spd[1];
 	};
 }
 #endif
